# openapi.yaml
...
openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dbaff4c29bdd95510990c0e
  title: TestowyProjekt
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  Debtor  -----
  /debtor/:
    post:
      x-skaffolder-id: 5dbb027d29bdd95510990da8
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dbb01cd29bdd95510990d82
      x-skaffolder-resource: Debtor
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Debtor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debtor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                Name:
                  type: string
                Pesel:
                  type: string
                Surname:
                  type: string
    get:
      x-skaffolder-id: 5dbb027729bdd95510990d9e
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dbb01cd29bdd95510990d82
      x-skaffolder-resource: Debtor
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Debtor
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Debtor"
  /debtor/{id}:
    delete:
      x-skaffolder-id: 5dbb027d29bdd95510990dad
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dbb01cd29bdd95510990d82
      x-skaffolder-resource: Debtor
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Debtor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dbb027d29bdd95510990da5
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dbb01cd29bdd95510990d82
      x-skaffolder-resource: Debtor
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Debtor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debtor"
    post:
      x-skaffolder-id: 5dbb027d29bdd95510990da7
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dbb01cd29bdd95510990d82
      x-skaffolder-resource: Debtor
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Debtor
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debtor"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Debtor"
    
  #-----  Document  -----
...
  /user/:
    post:
      x-skaffolder-id: 5dbaff4c29bdd95510990c1b
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dbaff4c29bdd95510990c22
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /user/{id}:
    delete:
      x-skaffolder-id: 5dbaff4c29bdd95510990c1e
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dbaff4c29bdd95510990c1c
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dbaff4c29bdd95510990c21
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dbaff4c29bdd95510990c18
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  VindicationCase  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
    Debtor:
      x-skaffolder-id: 5dbb01cd29bdd95510990d82
      x-skaffolder-id-db: 5dbaff4c29bdd95510990c0f
      x-skaffolder-id-entity: 5dbb01cd29bdd95510990d7e
      x-skaffolder-url: /debtor
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        Name:
          type: string
          x-skaffolder-id-attr: 5dbb01cd29bdd95510990d7f
          x-skaffolder-type: String
        Pesel:
          type: string
          x-skaffolder-id-attr: 5dbb01cd29bdd95510990d81
          x-skaffolder-type: String
        Surname:
          type: string
          x-skaffolder-id-attr: 5dbb01cd29bdd95510990d80
          x-skaffolder-type: String
      x-skaffolder-relations: 
      required:
        - _id
...
    User:
      x-skaffolder-id: 5dbaff4c29bdd95510990c18
      x-skaffolder-id-db: 5dbaff4c29bdd95510990c0f
      x-skaffolder-id-entity: 5dbaff4c29bdd95510990c10
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c13
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c14
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c12
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c16
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c15
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dbaff4c29bdd95510990c11
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...